Merged PR 141: Added an interactive interface to view and update barrier statuses, with design improvements and visual cues.
I built a display interface for the search task, which includes a box where the user can select the city, another box where the user can enter the checkpoint the user is searching for, and a card displaying the checkpoint the user searched for. This card contains all the important information, as well as a feedback card for each checkpoint, allowing the user to submit the checkpoint's status as soon as they see it.

Files I worked on:
DestinationSearch.js
DestinationSearch.css
App.js
navbar.js

Related work items: #538, #539